Move ./src/seedData under ./src/seeds for more consistency
