+++ backend/README.md
@@ -19,3 +19,6 @@ Backend project for Team Radio
 - Run `npm run doc` to generate documents
 - Open `doc/index.html` in the browser
 
+## Big Note
+
+- Please DO NOT write any ES6+ syntax in ./src/seeds. Knex JS runtime are not compiled by Babel. Writing any ES6+ syntax will make the application crash.
