+++ backend/src/models/index.js
@@ -1,6 +1,9 @@
 // authorization
 import authorization from './authorization';
 
+/**
+ * Importing models
+ */
 const models = {
   ...authorization,
 };
